[hw3] refactor for 3 layers structure
