+++ homework_3/package.json
@@ -8,7 +8,7 @@
     "build": "tsc",
     "dev-server": "nodemon --exec ts-node src/index.ts",
     "fulfill-table-users": "ts-node src/DB_utils/fulfill_DB_sequelize.ts",
-    "test-db-connect": "ts-node src/DB_utils/dbInit.ts"
+    "test-db-connect": "ts-node src/data-access/dbInit.ts"
   },
   "keywords": [
     "nodejs",
@@ -29,7 +29,6 @@
     "@types/hapi__joi": "^16.0.5",
     "@types/sequelize": "^4.28.8",
     "express": "^4.17.1",
-    "knex": "^0.20.8",
     "sequelize": "^5.21.3"
   }
 }
