--- homework_3/src/DB_connection_config_knex.json
@@ -1,9 +1,0 @@
-{
-  "client": "pg",
-  "connection": {
-    "host" : "localhost",
-    "user" : "postgres",
-    "password" : "dev",
-    "database" : "node_hero"
-  }
-}
